add Role.java, RoleRepository.java, UserRepository.java change User.java add implements UserDetails, change int id to long id, name to username, add password, passwordConfirm, roles, getAuthorities change UserServiceImpl add implements UserDetailsService, add UserRepository, setUserRepository, findByName, loadUserByUsername, mapRolesToAuthorities change WebSecurityConfig, add UserServiceImpl, setUserService, change userDetailsService to passwordEncoder and daoAuthenticationProvider

+++ src/main/java/ru/kata/spring/boot_security/demo/Service/UserService.java
@@ -1,19 +1,20 @@
 package ru.kata.spring.boot_security.demo.Service;
 
 
-
+import org.springframework.security.core.userdetails.UserDetailsService;
 import ru.kata.spring.boot_security.demo.models.User;
 
 import java.util.List;
-public interface UserService {
+public interface UserService  {
     void addUser(User user);
 
-    void removeUserById(int id);
+    void removeUserById(long id);
 
     List<User> getAllUsers();
 
-    User getUser(int id);
+    User getUser(long id);
 
-    void updateUser(int id, User user);
+    void updateUser(long id, User user);
+
 
 }
